Move YAML from `nodejs/yaml` into `yaml` so that it can be used outside Node
